Major bugfix
The assertion added earlier was not functioning when testing previous
commits (since assertions in `Haskell` only turn on when there are no
optimisations enabled). Thus a major bug creeped in. This commit fixes
this, and also fixes the testing code: now an incorrect result will lead
to an appropriate output.